+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `PcaBuilder` and `RandomizedPcaBuilder` to create instances of `Pca` and
+  `RandomizedPca`, respectively.
+
 ## [0.5.0] - 2021-05-02
 
 ### Changed